function layout is updated
